requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
Markdown==3.4.1
mypy-extensions==0.4.3
openpyxl==3.0.10
pathspec==0.10.3
platformdirs==2.6.2
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
